Autonomous opens up

+++ TeamCode/src/main/java/org/firstinspires/ftc/devices/ScissorLift.java
@@ -49,7 +49,7 @@ public class ScissorLift {
         liftMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
     }
 
-    public double getPosition () {
+    public int getPosition () {
         return liftMotor.getCurrentPosition();
     }
